Refactor move methods to individual services
